Reworked parse module + rename
