+++ src/sserver/parse/literal.py
@@ -1,379 +1,43 @@
-"""Definitions and functions for literal object handling."""
-
-
-import copy
-import inspect
-import operator
-from typing import Any, Dict, Optional, Tuple, Type, Union
-from sserver.parse import exception
-
-
-# Maps string operators onto the operator module functions
-_constant_operator_map = {
-    'True': {
-        'function': lambda: True,
-    },
-    'False': {
-        'function': lambda: False,
-    },
-    'None': {
-        'function': lambda: None,
-    },
-}
-
-_CONSTANT_OPERATOR_PRECEDENCE = 8
-
-_ARITHMETIC_OPERATOR_MAP = {
-    '+': {
-        'function': operator.add,
-        'precedence': 5,
-    },
-    '-': {
-        'function': operator.sub,
-        'precedence': 5,
-    },
-    '*': {
-        'function': operator.mul,
-        'precedence': 6,
-    },
-    '/': {
-        'function': operator.truediv,
-        'precedence': 6,
-    },
-    '//': {
-        'function': operator.floordiv,
-        'precedence': 6,
-    },
-    '%': {
-        'function': operator.mod,
-        'precedence': 6,
-    },
-    '**': {
-        'function': operator.pow,
-        'precedence': 7,
-    },
-}
-
-_LOGICAL_OPERATOR_MAP = {
-    '==': {
-        'function': operator.eq,
-    },
-    '!=': {
-        'function': operator.ne,
-    },
-    '>': {
-        'function': operator.gt,
-    },
-    '>=': {
-        'function': operator.ge,
-    },
-    '<': {
-        'function': operator.lt,
-    },
-    '<=': {
-        'function': operator.le,
-    },
-    'in': {
-        'function': lambda x, y: x in y,
-    }
-}
-
-_LOGICAL_OPERATOR_PRECEDENCE = 4
-
-_KEYWORD_OPERATOR_MAP = {
-    'is': {
-        'function': operator.is_,
-        'precedence': 4,
-    },
-    'not': {
-        'function': operator.not_,
-        'precedence': 3,
-    },
-    'and': {
-        'function': operator.and_,
-        'precedence': 2,
-    },
-    'or': {
-        'function': operator.or_,
-        'precedence': 1,
-    },
-}
-
-
-def _get_valid_operator_chars():
-    return (
-        tuple(_ARITHMETIC_OPERATOR_MAP.keys()) +
-        tuple(_constant_operator_map.keys()) +
-        tuple(_LOGICAL_OPERATOR_MAP.keys()) +
-        tuple(_KEYWORD_OPERATOR_MAP.keys())
-    )
-
-
-# Literal syntax map
-_literal_syntax_map = {}
-
-
-# Alias for literal match
-LiteralMatch = Dict[str, Union[str, Type, None]]
-
-# Alias for context
-Context = Dict[str, Any]
-
-
-# Getter for constant operator map
-def get_constant_operator_map() -> Dict[str, Dict[str, Any]]:
-    """The constant operator map.
-
-    Returns:
-        `Dict[str, Dict[str, Any]]`: The constant operator map.
-    """
-
-    return copy.deepcopy(_constant_operator_map)
-
-
-def add_constant_operator(operator_name: str, value: Any):
-    """Adds a constant operator to the operator map.
-
-    Args:
-        operator_name (`str`): The name of the operator.
-        value (`Any`): The value of the operator.
-    """
-
-    if operator_name in _constant_operator_map:
-        raise exception.OperatorAlreadyExistsException(
-            f'Operator "{operator_name}" already exists.'
-        )
-
-    _constant_operator_map[operator_name] = {
-        'function': lambda: value,
-    }
-
-
-# Add constant operators to the operator map
-def try_add_constant_operator(operator_name: str, value: Any
-                              ) -> bool:
-    """Attempts to add a constant operator to the operator map.
-
-    Args:
-        operator_name (`str`): The name of the operator.
-        value (`Any`): The value of the operator.
-
-    Returns:
-        `bool`: Whether the operator was added.
-    """
-
-    try:
-        add_constant_operator(operator_name, value)
-        return True
-
-    except exception.OperatorAlreadyExistsException:
-        return False
-
-
-# Getter for literal syntax map
-def get_literal_syntax_map() -> Dict[str, LiteralMatch]:
-    """The literal syntax map.
-
-    Returns:
-        `Dict[str, LiteralMatch]`: The literal syntax map.
-    """
-
-    return copy.deepcopy(_literal_syntax_map)
-
-
-def _register_literal_class(literal_class: Type,
-                            start_char: Union[str, Tuple[str]],
-                            value_type: Type,
-                            end_char: Optional[str],
-                            escape_char: Optional[str]):
-    """Registers a literal class to be used when parsing.
-
-    Args:
-        char (`str`): The character that starts the literal.
-        value_type (`Type`): The type of the literal value.
-        literal_class (`Type`): The literal class.
-        end_char (`Optional[str]`): The character that ends the literal.
-            Defaults to `None`.
-        escape_char (`Optional[str]`): The character that escapes the
-            end character. Defaults to `None`.
-
-    Raises:
-        `ValueError`: If the literal class passed is not a subclass
-            or `BaseLiteral`.
-        `LiteralCharacterAlreadyInUseException`: If the literal
-            character is already in use.
-    """
-
-    if not issubclass(literal_class, BaseLiteral):
-        raise ValueError(
-            f'Literal class {literal_class.__name__} must '
-            'be a subclass of BaseLiteral'
-        )
-
-    if start_char in _literal_syntax_map:
-        raise exception.LiteralCharacterAlreadyInUseException(
-            f'Character "{start_char}" is already in use'
-        )
-
-    _literal_syntax_map[start_char] = {
-        'value_type': value_type,
-        'escape_char': escape_char,
-        'end_char': end_char,
-        'literal_class': literal_class,
-    }
-
-
-# Decorator for literal classes to register them
-def register_literal_class(start_char: Union[str, Tuple[str]],
-                           value_type: Type,
-                           end_char: Optional[str] = None,
-                           escape_char: Optional[str] = None,
-                           ) -> callable:
-    """Decorator for registering a literal class to be used when
-        parsing.
-
-    Args:
-        char (`str`): The character that starts the literal.
-        value_type (`Type`): The type of the literal value.
-        end_char (`Optional[str]`): The character that ends the literal.
-            Defaults to `None`.
-        escape_char (`Optional[str]`): The character that escapes the
-            end character. Defaults to `None`.
-
-    Returns:
-        `callable`: The decorator function.
-
-    Raises:
-        `ValueError`: If start character and end character are not of
-            the same type, or if both start character and end
-            character are passed as tuples and have different lengths,
-            or if the start character is not a string or tuple of
-            strings.
-    """
-    def decorator(cls):
-        """The decorator function."""
-
-        if type(start_char) != type(end_char):
-            raise ValueError(
-                'char and end_char must be of the same type'
-            )
-
-        if isinstance(start_char, tuple):
-            if not isinstance(end_char, tuple):
-                raise ValueError(
-                    'end_char must be a tuple if char is a tuple and '
-                    'must have equal length'
-                )
-
-            if len(start_char) != len(end_char):
-                raise ValueError(
-                    'char and end_char must have equal length if '
-                    'either are passed as a tuple'
-                )
-
-            for character, end_character in zip(start_char, end_char):
-                _register_literal_class(
-                    cls,
-                    character,
-                    value_type,
-                    end_character,
-                    escape_char
-                )
-
-        elif not isinstance(start_char, str):
-            raise ValueError(
-                'char must be a string or a tuple of strings'
-            )
-
-        elif end_char is not None and not isinstance(end_char, str):
-            raise ValueError(
-                'end_char must be a string or a tuple of strings '
-                'and must be of the same type as char (with '
-                'equal length is char is a tuple)'
-            )
-
-        else:
-            _register_literal_class(
-                cls,
-                start_char,
-                value_type,
-                end_char,
-                escape_char
-            )
-
-        return cls
-
-    return decorator
-
-
-class Evaluatable:
-    """Represents an evaluatable object."""
-
-    def evaluate(self, context: Context) -> Any:
-        """Evaluates the object.
-
-        Args:
-            context (`Context`): The context to evaluate the object in.
-
-        Returns:
-            `Any`: The evaluated object.
-
-        Raises:
-            `NotImplementedError`: If the method is not implemented by
-                the subclass.
-        """
-
-        raise NotImplementedError((
-            'Evaluate not implemented for '
-            f'{self.__class__.__name__}'
-        ))
-
-
-class BaseLiteral(Evaluatable):
-    """Represents a literal value."""
-
-    def __init__(self, char: str, match:
-                 LiteralMatch):
-        """Initializes the literal value.
-
-        Args:
-            match (`LiteralMatch`): The match
-                dictionary from character map.
-
-        Raises:
-            `UnknownLiteralType`: If the literal type is unknown.
-        """
-
-        self._star_char = char
-        self._escape_char = match.get('escape_char')
-        self._end_char = match.get('end_char')
-        self._value_type = match.get('value_type')
-
-        if self._value_type is None:
-            raise exception.UnknownLiteralTypeException(
-                f'Unknown literal type: {self._value_type}'
-            )
-
-        self._value = ''
-
-    def __repr__(self) -> str:
-        """Returns the string representation of the literal value.
-
-        Returns:
-            `str`: The string representation of the literal value.
-        """
-
-        return f'<{self.__class__.__name__}, value: {self._value}>'
+"""Literal class declarations."""
+
+
+from typing import Any, List, Tuple
+from sserver.parse.parse import (
+    Expression,
+    parse_string_to_value,
+    exception,
+)
+from sserver.parse.base_literal import (
+    Evaluatable,
+    Operator,
+    BaseLiteral,
+    EnclosingLiteral,
+    NumericLiteral,
+    get_constant_operator_map,
+    add_constant_operator,
+    try_add_constant_operator,
+    get_literal_syntax_map,
+    is_unterminated_literal,
+    register_literal_class,
+    create_literal,
+    Context,
+    LiteralMatch,
+)
+
+
+@register_literal_class(
+    start_char=('"', '\''),
+    end_char=('"', '\''),
+    value_type=str,
+    escape_char='\\',
+)
+class StringLiteral(BaseLiteral):
+    """Represents a string literal value."""
 
     def _append_literal_match(self, char: str, position: int, match:
                               LiteralMatch
-                              ) -> bool:
+                              ):
         """Appends a literal match to the literal value.
-
-        Note:
-            Literal characters passed are considered to always
-                be consumed as they are either appended or are the
-                end character.
 
         Args:
             char (`str`): The character to append.
@@ -382,17 +46,43 @@ class BaseLiteral(Evaluatable):
 
         Returns:
             `bool`: Whether the literal value is complete.
-
-        Raises:
-            `ExpressionSyntaxException`: If a literal character is
-                matched at an unexpected position.
         """
 
-        # Raise a syntax error if append_literal_match is not
-        # overridden
-        raise exception.ExpressionSyntaxException(
-            f'Unexpected literal match: {char} at position {position}'
-        )
+        if self._is_end_char(char):
+            return True
+
+        self._try_strip_escape_char(char)
+        self._value += char
+
+        return False
+
+
+@register_literal_class(
+    start_char='[',
+    end_char=']',
+    value_type=list
+)
+class ListLiteral(EnclosingLiteral):
+    """Represents a list literal value."""
+
+    def __init__(self, char: str, match:
+                 LiteralMatch):
+
+        super().__init__(char, match)
+
+        # Initialize the list value
+        self._value = [None]
+
+    def __repr__(self) -> str:
+        output = f'<{self.__class__.__name__}, value: ['
+
+        for value in self._value:
+            if value is not None:
+                output += f'{value}, '
+
+        output = output[:-2] + ']>'
+
+        return output
 
     def _append_character(self, char: str, position: int
                           ) -> Tuple[bool, bool]:
@@ -410,407 +100,119 @@ class BaseLiteral(Evaluatable):
                 consumed by the literal value.
         """
 
-        if self._is_end_char(char):
+        # Check if the character is a list separator or whitespace
+        if char == ',':
+            self._value.append(None)
+            return False, True
+
+        previous_value = self._value[-1]
+
+        # Only ignore whitespace if it comes after a list separator
+        # or the start of the list
+        if char.isspace() and previous_value is None:
+            return False, True
+
+        if self._char_terminates_literal(char, previous_value):
+            # If the last item in the list is None, remove it
+            if previous_value is None:
+                self._value.pop()
+
             return True, True
 
-        self._try_strip_escape_char(char)
+        # If all the above passed, add the character to the current
+        # list item
+        if previous_value is None:
+            self._value[-1] = char
+
+        else:
+            self._value[-1] += char
+
+        return False, True
+
+    def evaluate(self, context: Context) -> List[Any]:
+        """Evaluates the list literal value.
+
+        Args:
+            context (`Context`): The context dictionary.
+
+        Returns:
+            `List[Any]`: The evaluated list literal value.
+
+        Raises:
+            `ExpressionSyntaxException`: If the list literal
+                encounters an unexpected separator character.
+        """
+
+        # Evaluate the list by parsing the strings of each item
+        evaluated_list = []
+        for index, value in enumerate(self._value):
+            # If the value is None, a syntax error has occurred
+            if value is None:
+                # Generate list string for error message and replace
+                # None with ""
+
+                exception_list_string = str(self._value).replace(
+                    'None',
+                    ''
+                )
+
+                raise exception.ExpressionSyntaxException(
+                    f'Unexpected list separator between list index '
+                    f'{index - 1} and list index {index} in list: '
+                    f'{exception_list_string}'
+                )
+
+            # Parse the item string
+            parsed_value = parse_string_to_value(context, value)
+
+            evaluated_list.append(parsed_value)
+
+        return evaluated_list
+
+
+@register_literal_class(
+    start_char='(',
+    end_char=')',
+    value_type=Expression
+)
+class ParenthesisLiteral(EnclosingLiteral):
+    """Represents a parenthesized literal."""
+
+    def _append_character(self, char: str, position: int
+                          ) -> Tuple[bool, bool]:
+        # Check if the character terminates the literal
+        # Add an additional "(" to the value to account for the
+        # open parenthesis that is not included in the value
+        if self._char_terminates_literal(char, self._value):
+            return True, True
+
+        # Append the character normally
         self._value += char
 
         return False, True
 
-    def _previous_char_was_escape(self) -> bool:
-        """Returns whether the previous character was an escape
-        character.
-
-        Returns:
-            `bool`: Whether the previous character was an escape
-                character.
-        """
-
-        # By default, handle appending characters as a string
-        previous_char_was_escape = False
-
-        if len(self._value) > 0:
-            previous_char_was_escape = (
-                self._value[-1] == self._escape_char
-            )
-
-        return previous_char_was_escape
-
-    def _is_end_char(self, char: str):
-        """Checks if `char` is a valid end to the literal.
-
-        Args:
-            char (`str`): The character to check.
-
-        Returns:
-            `bool`: Whether the character is a valid end to the
-                literal.
-        """
-
-        return (
-            char == self._end_char
-            and not self._previous_char_was_escape()
-        )
-
-    def _try_strip_escape_char(self, char: str):
-        """Strips the escape character from a character if it is
-        present.
-
-        Args:
-            char (`str`): The current character.
-        """
-
-        # By default, handle appending characters as a string
-        previous_char_was_escape = self._previous_char_was_escape()
-
-        is_end_char = char == self._end_char
-
-        # Strip the escape character from the value if the end
-        # character is met after an escape character
-        if is_end_char and previous_char_was_escape:
-            self._value = self._value[:-1]
-
-    def _can_terminate(self) -> bool:
-        """Returns whether the literal value can be terminated.
-
-        Note:
-            This method determines if the literal value can be
-                immediately terminated.
-
-        Returns:
-            `bool`: Whether the literal value can be terminated.
-        """
-
-        return self._end_char is None
-
     def evaluate(self, context: Context) -> Any:
-        """Evaluates the literal value.
-
-        Args:
-            context (`Context`): The context to evaluate the
-                literal value in.
-
-        Returns:
-            `Any`: The evaluated literal value.
-
-        Raises:
-            `ExpressionSyntaxException`: If the literal value failed
-                to be cast to its value type.
-        """
-
-        # If value_type is a tuple, try each type and the first that
-        # succeeds is the value type
-        if isinstance(self._value_type, tuple):
-            for value_type in self._value_type:
-                try:
-                    return value_type(self._value)
-                except ValueError:
-                    pass
-
-            raise exception.ExpressionSyntaxException(
-                f'Invalid literal value: {self._value}'
-            )
-
-        return self._value_type(self._value)
-
-
-class Operator:
-    """Represents an arithmetic or logical operator."""
-
-    def __init__(self, char: str):
-        """Initializes the operator.
-
-        Raises:
-            `UnknownOperatorException`: If the operator is unknown.
-        """
-
-        if char not in _get_valid_operator_chars():
-            raise exception.UnknownOperatorException(
-                f'Unknown operator: {char}'
-            )
-
-        self._operator = char
-
-    def __repr__(self) -> str:
-        return (
-            f'<{self.__class__.__name__}, '
-            f'operator: "{self._operator}">'
-        )
-
-    def __call__(self, *args):
-        """Calls the operator function."""
-
-        func = self._get_function()
-
-        return func(*args)
-
-    @property
-    def precedence(self) -> int:
-        """Gets the operators precedence.
-
-        Returns:
-            `int`: The operators precedence.
-
-        Raises:
-            `UnknownOperatorException`: If the operator is unknown.
-        """
-
-        if self._operator in _LOGICAL_OPERATOR_MAP:
-            return _LOGICAL_OPERATOR_PRECEDENCE
-
-        elif self._operator in _constant_operator_map:
-            return _CONSTANT_OPERATOR_PRECEDENCE
-
-        match = self._get_operator_map_match()
-
-        precedence = match.get('precedence')
-
-        if precedence is None:
-            raise exception.MissingOperatorPrecedenceException(
-                f'Missing operator precedence: {self._operator}'
-            )
-
-        return precedence
-
-    @property
-    def argument_count(self) -> int:
-        """Gets the number of arguments the operator takes.
-
-        Returns:
-            `int`: The number of arguments the operator takes.
-        """
-
-        func = self._get_function()
-
-        return len(inspect.signature(func).parameters)
-
-    @classmethod
-    def is_valid_operator(cls, char: str) -> bool:
-        """Checks if the character is a valid operator.
-
-        Args:
-            char (`str`): The character to check.
-
-        Returns:
-            `bool`: True if the character is a valid operator, False
-                if not.
-        """
-
-        return char in _get_valid_operator_chars()
-
-    @classmethod
-    def string_could_be_operator(cls, string: str) -> bool:
-        """Checks if the string could be an operator.
-
-        Args:
-            string (`str`): The string to check.
-
-        Returns:
-            `bool`: True if the string could be an operator, False if
-                not.
-        """
-
-        return any(
-            operator.startswith(string)
-            for operator in _get_valid_operator_chars()
-        )
-
-    def _get_operator_map_match(self) -> LiteralMatch:
-        """Gets the respective operators match.
-
-        Returns:
-            `LiteralMatch`: The operators respecive match.
-
-        Raises:
-            `UnknownOperatorException`: If the operator is unknown.
-        """
-
-        match = None
-
-        if self._operator in _ARITHMETIC_OPERATOR_MAP:
-            match = _ARITHMETIC_OPERATOR_MAP[self._operator]
-
-        elif self._operator in _constant_operator_map:
-            match = _constant_operator_map[self._operator]
-
-        elif self._operator in _LOGICAL_OPERATOR_MAP:
-            match = _LOGICAL_OPERATOR_MAP[self._operator]
-
-        elif self._operator in _KEYWORD_OPERATOR_MAP:
-            match = _KEYWORD_OPERATOR_MAP[self._operator]
-
-        # Check a match was found
-        if match is None:
-            raise exception.UnknownOperatorException(
-                f'Unknown operator: {self._operator}'
-            )
-
-        return match
-
-    def _get_function(self) -> callable:
-        """Gets the operator function.
-
-        Returns:
-            `callable`: The operator function.
-        """
-
-        match = self._get_operator_map_match()
-        func = match.get('function')
-
-        if func is None:
-            raise exception.MissingOperatorFunctionException(
-                f'Missing operator function: {self._operator}'
-            )
-
-        return func
-
-
-def is_unterminated_literal(value: str) -> bool:
-    """Returns whether the value is an unterminated literal.
-
-    Args:
-        value (`str`): The string value to check.
-
-    Returns:
-        `bool`: Whether the value is unterminated.
-
-    Raises:
-        `ExpressionSyntaxException`: If a literal close character is
-            encountered without an open character.
-    """
-
-    # If one or less characters present, the value cannot be
-    # terminated
-    if len(value) <= 1:
-        return False
-
-    # Keep track of all open literals
-    # Create mappings for open and close characters onto the manifest,
-    # taking advantage of dict mutability
-    literal_character_manifest = {
-        char: {
-            'match': match,
-            'open_count': 0,
-        }
-        for char, match in _literal_syntax_map.items()
-    }
-
-    literal_open_chars = {
-        char: literal_character_manifest[char]
-        for char, match in _literal_syntax_map.items()
-    }
-
-    literal_end_chars = {
-        match['end_char']: literal_open_chars[char]
-        for char, match in _literal_syntax_map.items()
-        if match['end_char'] is not None
-    }
-
-    masking_end_char = None
-    is_masking = False
-
-    value_clone = value
-
-    # First remove all escaped values
-    for char, manifest_value in literal_character_manifest.items():
-        match = manifest_value['match']
-        end_char = match['end_char']
-        escape_char = match['escape_char']
-
-        if escape_char is not None:
-            # Remove all escape open and end chars
-            value_clone = value_clone.replace(
-                escape_char + char,
-                ''
-            )
-
-            if end_char is not None:
-                value_clone = value_clone.replace(
-                    escape_char + end_char,
-                    ''
-                )
-
-    for index, char in enumerate(value_clone):
-        # First check whether the current literal is masking,
-        # only caring if the masking is finished
-        if is_masking:
-            if char == masking_end_char:
-                is_masking = False
-                masking_end_char = None
-
-                literal_end_chars[char]['open_count'] -= 1
-
-            continue
-
-        # Check if the character is an open character
-        if char in literal_open_chars:
-            current_match = literal_open_chars[char]['match']
-
-            # Check whether the open and close characters are the same
-            if current_match['end_char'] == char:
-                # If the open and close characters are the same,
-                # a new masking literal has been started as already
-                # open masking literals would be checked earlier
-                is_masking = True
-                masking_end_char = char
-
-            literal_open_chars[char]['open_count'] += 1
-
-            continue
-
-        # Check if the character is an end character
-        if char in literal_end_chars:
-            # Check if the literal is open
-            if literal_end_chars[char]['open_count'] > 0:
-                literal_end_chars[char]['open_count'] -= 1
-
-            else:
-                # If no open literal is found, a syntax error has occurred
-                raise exception.ExpressionSyntaxException(
-                    f'Unexpected literal close character: {char} '
-                    f'in {value} near index {index}'
-                )
-
-            continue
-
-    # Finally, check if any literals are still open
-    for char, manifest in literal_character_manifest.items():
-        if manifest['open_count'] > 0:
-            return True
-
-    return False
-
-
-def create_literal(char: str, match: LiteralMatch) -> BaseLiteral:
-    """Creates a new literal value.
-
-    Args:
-        char (`str`): The character that started the literal.
-        match (`LiteralMatch`): The match
-            dictionary from character map.
-
-    Returns:
-        `BaseLiteral`: The new literal value.
-
-    Raises:
-        `ValueError`: If a value type that is not a type is passed.
-        `UnknownLiteralType`: If the literal type is unknown.
-    """
-
-    # Use value type to determine which literal to create
-    value_type = match.get('value_type')
-    literal_class = match.get('literal_class')
-
-    if not isinstance(value_type, type):
-        raise ValueError((
-            f'Invalid value type for literal: char="{char}", '
-            f'value_type="{value_type}"'
-        ))
-
-    if literal_class is None:
-        raise exception.UnknownLiteralTypeException(
-            f'Unknown literal type: {value_type}'
-        )
-
-    return literal_class(char, match)
+        return parse_string_to_value(context, self._value)
+
+# Expose literal classes and functions
+__all__ = [
+    'Expression',
+    'Evaluatable',
+    'Operator',
+    'BaseLiteral',
+    'EnclosingLiteral',
+    'NumericLiteral',
+    'StringLiteral',
+    'ListLiteral',
+    'ParenthesisLiteral',
+    'parse_string_to_value',
+    'get_constant_operator_map',
+    'add_constant_operator',
+    'try_add_constant_operator',
+    'get_literal_syntax_map',
+    'is_unterminated_literal',
+    'register_literal_class',
+    'create_literal',
+    'Context',
+    'LiteralMatch',
+]
